Fix error with checking verification
